Resolver, Query and Schema of UpdateTruckDay done
